Add: modal content to be dynamically served by javascript
