--- src/js/dom-triggers.js
@@ -1,8 +1,0 @@
-const popupModal = document.querySelector('.popup-modal');
-const btnOpenPopupModal = document.querySelector('.btn-open-popup-modal');
-const btnClosePopupModal = document.querySelector('.btn-close-popup-modal');
-
-export const modalTrigger = () => {
-  btnOpenPopupModal.addEventListener('click', () => popupModal.classList.remove('hidden'));
-  btnClosePopupModal.addEventListener('click', () => popupModal.classList.add('hidden'));
-};
